requirements.txt
# Media Studio - Python Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1

# Image Processing
pillow>=10.2.0
numpy>=1.24.0

# Hebrew/Arabic RTL Text Support
python-bidi>=0.4.2
arabic-reshaper>=3.0.0

# AI Integration
google-generativeai>=0.3.2

# Environment
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
